clean a bit the main
